# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo App Api Docs",
        "description": "This is official docs for the Todo App",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Local server"
        },
        {
            "url": "https://laravel-todo.me/api/",
            "description": "Production server"
        }
    ],
    "paths": {
        "/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get all notes",
                "description": "Fetches a paginated list of notes.",
                "operationId": "9393a3bea0b52c3edef159a70977cbbe",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of notes per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch notes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a specific note",
                "description": "Fetches a specific note by its ID.",
                "operationId": "4ff1db70f3fbc8ab14e315c81fbff433",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update a note",
                "description": "Updates the details of a specific note.",
                "operationId": "29a9163f38861fad93ae0510c516c5cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Note"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description."
                                    },
                                    "is_completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Note updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update note",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete a note",
                "description": "Deletes a specific note.",
                "operationId": "193c1092c012659cba45810e88e7a619",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete note",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/record_lists/{record_list_id}/notes": {
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create a new note",
                "description": "Creates a new note under a specific record list.",
                "operationId": "2fdfb12cf86822ae459b87185b0d11b4",
                "parameters": [
                    {
                        "name": "record_list_id",
                        "in": "path",
                        "description": "ID of the record list",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Note"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a note."
                                    },
                                    "is_completed": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create note",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FailureResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FailureResponse": {
                "title": "Failure Response",
                "description": "Standard failure response structure",
                "properties": {
                    "success": {
                        "description": "Indicates if the request failed",
                        "type": "boolean",
                        "default": false,
                        "example": false
                    },
                    "message": {
                        "description": "Message describing the failure",
                        "type": "string",
                        "example": "Operation failed"
                    },
                    "errors": {
                        "description": "Any errors or data related to the failure",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "Success Response",
                "description": "Standard success response structure",
                "properties": {
                    "success": {
                        "description": "Indicates if the request was successful",
                        "type": "boolean",
                        "default": true,
                        "example": true
                    },
                    "message": {
                        "description": "Message describing the outcome of the operation",
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "description": "The actual data returned from the request",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}